including CachingRasterRDDReader for S3
