+++ spark/src/main/scala/geotrellis/spark/io/s3/CachingRasterRDDReader.scala
@@ -1,0 +1,101 @@
+package geotrellis.spark.io.s3
+
+import java.io.{FileOutputStream, FileInputStream, File}
+import java.util.concurrent.Executors
+
+import com.amazonaws.services.s3.model.AmazonS3Exception
+import geotrellis.spark._
+import geotrellis.spark.io.index.KeyIndex
+import org.apache.spark.SparkContext
+import com.typesafe.scalalogging.slf4j._
+import scala.reflect.ClassTag
+import geotrellis.spark.io.avro._
+import org.apache.avro._
+import spray.json.DefaultJsonProtocol._
+import spray.json._
+
+
+import scala.util.Try
+import scalaz.concurrent.Task
+import scalaz.stream.Process
+
+class CachingRasterRDDReader[K: AvroRecordCodec: Boundable: ClassTag](cacheDirectory: File) extends  RasterRDDReader[K]  with LazyLogging {
+  /** Converting lower bound of Range to first Key for Marker */
+  override val indexToPath: (Long, Int) => String = encodeIndex
+
+  logger.info(s"Caching S3 reads to $cacheDirectory")
+
+  override def read(
+    attributes: S3AttributeStore,
+    s3client: ()=>S3Client,
+    layerMetaData: S3LayerMetaData,
+    keyBounds: KeyBounds[K],
+    keyIndex: KeyIndex[K],
+    numPartitions: Int)
+  (layerId: LayerId, queryKeyBounds: Seq[KeyBounds[K]])
+  (implicit sc: SparkContext): RasterRDD[K] = {
+    val bucket = layerMetaData.bucket
+    val dir = layerMetaData.key
+    val rasterMetaData = layerMetaData.rasterMetaData
+    
+    // TODO: now that we don't have to LIST, can we prefix key with a hash to get better throughput ?
+    val ranges = queryKeyBounds.flatMap(keyIndex.indexRanges(_))
+    val bins = RasterRDDReader.balancedBin(ranges, numPartitions)
+    logger.debug(s"Loading layer from $bucket $dir, ${ranges.length} ranges split into ${bins.length} bins")
+
+    // Broadcast the functions and objects we can use in the closure
+    val writerSchema: Schema = (new Schema.Parser).parse(attributes.read[JsObject](layerId, "schema").toString())
+
+    val maxWidth = maxIndexWidth(keyIndex.toIndex(keyBounds.maxKey))
+    val BC = sc.broadcast((
+      s3client,
+      { (index: Long) => indexToPath(index, maxWidth) },
+      { (key: K) => queryKeyBounds.includeKey(key) },
+      geotrellis.spark.io.avro.recordCodec(implicitly[AvroRecordCodec[K]],
+        geotrellis.spark.io.avro.tileUnionCodec),
+      writerSchema,
+      cacheDirectory,
+      layerId
+    ))
+
+    val rdd =
+      sc
+        .parallelize(bins, bins.size)
+        .mapPartitions { rangeList =>
+          val (fS3client, toPath, includeKey, recCodec, schema, cacheDir, id) = BC.value
+          val s3client = fS3client()
+
+          rangeList
+            .flatMap(ranges => ranges)
+            .flatMap { range =>
+              {for (index <- range._1 to range._2) yield {
+                val path = List(dir, toPath(index)).filter(_.nonEmpty).mkString("/")
+                val cachePath = new File(cacheDir, s"${id.name}__${id.zoom}__$index")
+                
+                Try{
+                  new FileInputStream(cachePath)
+                }
+                .recover {
+                  case e => s3client.getObject(bucket, path).getObjectContent
+                }
+                .map { is =>
+                  val bytes = org.apache.commons.io.IOUtils.toByteArray(is)
+
+                  val cacheOut = new FileOutputStream(cachePath)
+                  cacheOut.write(bytes)
+                  cacheOut.close()
+
+                  val recs = AvroEncoder.fromBinary(schema, bytes)(recCodec)
+                  recs.filter { row => includeKey(row._1) }
+                }
+                .recover {
+                  case e: AmazonS3Exception if e.getStatusCode == 404 => Seq.empty
+                }.get
+
+              }}.flatten
+            }
+        }
+
+    new RasterRDD[K](rdd, rasterMetaData)
+  }
+}
