+++ spark/src/main/scala/geotrellis/spark/io/s3/package.scala
@@ -4,11 +4,11 @@ import geotrellis.spark._
 import geotrellis.spark.io.avro.KeyCodecs._
 
 package object s3 {
-  implicit def s3SpatialRasterRDDReader = new RasterRDDReader[SpatialKey]
+  implicit def s3SpatialRasterRDDReader = new DirectRasterRDDReader[SpatialKey]
   implicit def s3SpatialRasterRDDWriter = new RasterRDDWriter[SpatialKey]
   implicit def s3SpatialRasterTileReader = new TileReader[SpatialKey]
 
-  implicit def s3SpaceTimeRasterRDDReader = new RasterRDDReader[SpaceTimeKey]
+  implicit def s3SpaceTimeRasterRDDReader = new DirectRasterRDDReader[SpaceTimeKey]
   implicit def s3SpaceTimeRasterRDDWriter = new RasterRDDWriter[SpaceTimeKey]
   implicit def s3SpaceTimeRasterTileReader = new TileReader[SpaceTimeKey]
